agri-identity-platform/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
cryptography
python-multipart
python-dotenv
email-validator
